Added initial skeleton for the application
The application is separated into a simulation backend and ideally any c++ game library or custom renderer
can be used to interface with the simulation. Right now have it set up to work with SFML.

include/ has some helper .hpp files for a N dimension Vector struct and a Particle struct that uses
the Vector implementation. There is also a Renderer abstract base class that the specfic implementations
can inherit from.